Correct sensors data by floatField

+++ backend/track_my_moves/models/sensors_interval.py
@@ -4,12 +4,12 @@ from django.db import models
 class SensorsInterval(models.Model):
     interval = models.ForeignKey("ActivityInterval", on_delete=models.CASCADE)
     time = models.PositiveIntegerField()
-    accel_x = models.IntegerField()
-    accel_y = models.IntegerField()
-    accel_z = models.IntegerField()
-    gyros_x = models.IntegerField()
-    gyros_y = models.IntegerField()
-    gyros_z = models.IntegerField()
+    accel_x = models.FloatField()
+    accel_y = models.FloatField()
+    accel_z = models.FloatField()
+    gyros_x = models.FloatField()
+    gyros_y = models.FloatField()
+    gyros_z = models.FloatField()
     
     def save(self, *args, **kwargs):
         self.full_clean()
